@@ -1,45 +1,31 @@
 [build-system]
-requires = ["setuptools>=64", "setuptools_scm[toml]>=6.2", "wheel"]
+requires = ["setuptools>=64", "setuptools_scm[toml]>=8"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "i10-bluesky"
+name = "i10_bluesky"
 classifiers = [
     "Development Status :: 3 - Alpha",
     "License :: OSI Approved :: Apache Software License",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
 ]
-description = ""
-dependencies = [
-    "typing-extensions;python_version<'3.8'",
-
-    "pyqt5",
-    "setuptools",
-    "networkx>=2.0",
-    "numpy",
-    "packaging",
-    "pint",
-    "event-model",
-    "p4p",
-    "pyyaml",
-] # Add project dependencies here, e.g. ["click", "numpy"]
+description = "i10 bluesky project"
+dependencies = [] # Add project dependencies here, e.g. ["click", "numpy"]
 dynamic = ["version"]
-#version = "0.0.1"         #need to make this non manual but dynamic version currently include local version seed and it not the statand for release. 
 license.file = "LICENSE"
-readme = "README.rst"
-requires-python = ">=3.9"
+readme = "README.md"
+requires-python = ">=3.10"
 
 [project.optional-dependencies]
 dev = [
-    "bluesky",
-    "ophyd-async",
-    "black",
-    "mypy",
+    "copier",
+    "myst-parser",
     "pipdeptree",
     "pre-commit",
     "pydata-sphinx-theme>=0.12",
+    "pyright",
     "pytest",
     "pytest-cov",
     "ruff",
